Removed superfluous interfaces and dead code
