+++ src/main/java/uk/gov/ons/ctp/response/sample/message/PartyReceiver.java
@@ -1,14 +1,21 @@
 package uk.gov.ons.ctp.response.sample.message;
 
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.integration.annotation.MessageEndpoint;
+import org.springframework.integration.annotation.ServiceActivator;
 import uk.gov.ons.ctp.response.party.definition.PartyCreationRequestDTO;
+import uk.gov.ons.ctp.response.sample.service.SampleService;
 
-/** The reader of Party(s) from queue */
-public interface PartyReceiver {
+@Slf4j
+@MessageEndpoint
+public class PartyReceiver {
 
-  /**
-   * To process a Party read from queue
-   *
-   * @param party the java representation of the message body
-   */
-  void acceptParty(PartyCreationRequestDTO party) throws Exception;
+  @Autowired private SampleService sampleService;
+
+  @ServiceActivator(inputChannel = "partyTransformed", adviceChain = "partyRetryAdvice")
+  public void acceptParty(PartyCreationRequestDTO party) throws Exception {
+    log.debug("acceptParty {}", party);
+    sampleService.sendToPartyService(party);
+  }
 }
