+++ src/main/java/uk/gov/ons/ctp/response/sample/message/SampleOutboundPublisher.java
@@ -1,11 +1,47 @@
 package uk.gov.ons.ctp.response.sample.message;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.amqp.rabbit.core.RabbitTemplate;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.integration.annotation.MessageEndpoint;
 import uk.gov.ons.ctp.common.error.CTPException;
 import uk.gov.ons.ctp.response.sample.domain.model.SampleSummary;
 
-/** An interface for publishing notifications that a sample has been uploaded */
-public interface SampleOutboundPublisher {
-  void sampleUploadStarted(SampleSummary sampleSummary) throws CTPException;
+/** Publisher of messages about samples uploaded */
+@Slf4j
+@MessageEndpoint
+public class SampleOutboundPublisher {
 
-  void sampleUploadFinished(SampleSummary sampleSummary) throws CTPException;
+  private static final String UPLOAD_STARTED_ROUTING_KEY = "Sample.SampleUploadStarted.binding";
+  private static final String UPLOAD_FINISHED_ROUTING_KEY = "Sample.SampleUploadFinished.binding";
+
+  @Qualifier("sampleOutboundRabbitTemplate")
+  @Autowired
+  private RabbitTemplate rabbitTemplate;
+
+  @Autowired private ObjectMapper objectMapper;
+
+  public void sampleUploadStarted(SampleSummary sampleSummary) throws CTPException {
+    sendMessage(UPLOAD_STARTED_ROUTING_KEY, sampleSummary);
+  }
+
+  public void sampleUploadFinished(SampleSummary sampleSummary) throws CTPException {
+    sendMessage(UPLOAD_FINISHED_ROUTING_KEY, sampleSummary);
+  }
+
+  private void sendMessage(String routingKey, SampleSummary sampleSummary) throws CTPException {
+    try {
+      String message = this.objectMapper.writeValueAsString(sampleSummary);
+      log.debug("Sending message to routing key {} - {}", routingKey, message);
+      rabbitTemplate.convertAndSend(routingKey, message);
+    } catch (JsonProcessingException e) {
+      throw new CTPException(
+          CTPException.Fault.SYSTEM_ERROR,
+          String.format("Failed to create JSON message from sample summary - %v", sampleSummary),
+          e);
+    }
+  }
 }
