--- src/test/java/uk/gov/ons/ctp/response/sample/message/SFTPSampleReceiverImplTest.java
@@ -1,5 +1,0 @@
-package uk.gov.ons.ctp.response.sample.message;
-
-/** Created by wardlk on 04/07/2017. */
-// @RunWith(MockitoJUnitRunner.class)
-public class SFTPSampleReceiverImplTest {}
