+++ src/test/resources/uk/gov/ons/ctp/response/sample/service/SampleServiceTest.SampleSummary.json
