+++ src/test/resources/uk/gov/ons/ctp/response/sample/service/business-survey-sample.xml
