+++ src/test/resources/uk/gov/ons/ctp/response/sample/service/census-survey-sample.xml
