+++ src/test/resources/uk/gov/ons/ctp/response/sample/service/social-survey-sample.xml
